arranged and renamed some files
i have to arrange and rename them kay kinda confusing na sya. i hope u understand huhu
