+++ lib/screens/customer_screen/customer_list/customerlist_widget.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+//import 'package:provider/provider.dart';
+
+class CustomerListWidget extends StatefulWidget {
+  const CustomerListWidget({ Key? key }) : super(key: key);
+
+  @override
+  State<CustomerListWidget> createState() => _CustomerListWidgetState();
+}
+
+class _CustomerListWidgetState extends State<CustomerListWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      
+    );
+  }
+}
