+++ lib/screens/login_screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:dennis_lechon_crm/widgets/loading.dart';
 import 'package:dennis_lechon_crm/screens/homebuttons/calendar_screen.dart';
-import 'package:dennis_lechon_crm/screens/homebuttons/customerlist_screen.dart';
-import 'package:dennis_lechon_crm/screens/homebuttons/orderlist_screen.dart';
+import 'package:dennis_lechon_crm/screens/homebuttons/customer_screen.dart';
+import 'package:dennis_lechon_crm/screens/homebuttons/order_screen.dart';
 import 'package:dennis_lechon_crm/screens/login_screens/signin_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
